+++ src/Scalar2DRepresentation.h
@@ -7,13 +7,13 @@
 #include <iostream>
 
 class Scalar2D {
-    int x_, y_;
+    double x_, y_;
 public:
-    Scalar2D( int x, int y ) : x_( x ), y_( y ) {}
-    int getX() const { return x_; }
-    int getY() const { return y_; }
-    virtual void setX( int new_x ) { x_ = new_x; }
-    virtual void setY( int new_y ) { y_ = new_y; }
+    Scalar2D( double x, double y ) : x_( x ), y_( y ) {}
+    double getX() const { return x_; }
+    double getY() const { return y_; }
+    virtual void setX( double new_x ) { x_ = new_x; }
+    virtual void setY( double new_y ) { y_ = new_y; }
     virtual std::string toString() const = 0;
 };
 
@@ -31,10 +31,10 @@ public:
 //@todo rename to Coordinates
 class Point : public Scalar2D {
 public:
-    Point( unsigned int x, unsigned int y ) : Scalar2D(x, y) {}
+    Point( double x, double y ) : Scalar2D(x, y) {}
     //Point coordinates can't be negative
-    virtual void setX( int new_x );
-    virtual void setY( int new_y );
+    virtual void setX( double new_x );
+    virtual void setY( double new_y );
     virtual std::string toString() const;
     virtual Point getMidpoint( const Point & p) const;
     virtual Direction operator-( const Point & p ) const { return Direction(getX() - p.getX(),  getY() - p.getY() ); }
@@ -48,8 +48,8 @@ class Resolution : public Scalar2D {
 public:
     Resolution( unsigned int x, unsigned int y ) : Scalar2D(x, y) {}
     //Resolution can't be negative
-    virtual void setX( int new_x );
-    virtual void setY( int new_y );
+    virtual void setX( double new_x );
+    virtual void setY( double new_y );
     virtual std::string toString() const;
 };
